Reconstruct surface now as other methods exposed

+++ pysplashsurf/tests/test_calling.py
@@ -3,7 +3,7 @@ import numpy as np
 import meshio
 
 def test_reconstruct_surface():
-    particles = np.array(meshio.read("./ParticleData_Fluid_5.vtk").points, dtype=np.float32)
+    particles = np.array(meshio.read("./ParticleData_Fluid_5.vtk").points, dtype=np.float64)
     reconstruction = pysplashsurf.reconstruct_surface(particles, enable_multi_threading=True, particle_radius=0.025, 
                                                                  rest_density=1000.0, smoothing_length=2.0, cube_size=0.5, 
                                                                  iso_surface_threshold=0.6, aabb_min=np.array([0.0, 0.0, 0.0]), aabb_max=np.array([2.0, 2.0, 2.0]))
@@ -18,7 +18,7 @@ def test_reconstruct_surface():
 test_reconstruct_surface()
 
 def test_post_processing():
-    particles = np.array(meshio.read("./ParticleData_Fluid_5.vtk").points, dtype=np.float64)
+    particles = np.array(meshio.read("./ParticleData_Fluid_5.vtk").points, dtype=np.float32)
     reconstruction = pysplashsurf.reconstruct_surface(particles, enable_multi_threading=True, particle_radius=0.025, 
                                                                  rest_density=1000.0, smoothing_length=2.0, cube_size=0.5, 
                                                                  iso_surface_threshold=0.6)
@@ -29,7 +29,7 @@ def test_post_processing():
 test_post_processing()
 
 def test_marching_cubes_cleanup(): 
-    particles = np.array(meshio.read("./ParticleData_Fluid_5.vtk").points, dtype=np.float64)
+    particles = np.array(meshio.read("./ParticleData_Fluid_5.vtk").points, dtype=np.float32)
     reconstruction = pysplashsurf.reconstruct_surface(particles, enable_multi_threading=True, particle_radius=0.025, 
                                                                  rest_density=1000.0, smoothing_length=2.0, cube_size=0.5, 
                                                                  iso_surface_threshold=0.6, aabb_min=np.array([0.0, 0.0, 0.0]), aabb_max=np.array([2.0, 2.0, 2.0]))
